requirements.txt
2captcha-python==1.2.4
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.16.0
selenium-stealth==1.0.6
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
undetected-chromedriver==3.5.5
urllib3==2.2.1
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
zenrows==1.3.2
psycopg==3.1.18
psycopg-binary==3.1.18
